Fix fdbserver/core link test
