+++ fdbserver/backupworker/BackupWorker.cpp
@@ -24,10 +24,10 @@
 #include "fdbclient/CommitProxyInterface.h"
 #include "fdbclient/SystemData.h"
 #include "fdbserver/core/BackupInterface.h"
-#include "fdbserver/core/BackupProgress.h"
+#include "fdbserver/logsystem/BackupProgress.h"
 #include "fdbserver/core/Knobs.h"
 #include "fdbserver/core/LogProtocolMessage.h"
-#include "fdbserver/core/LogSystem.h"
+#include "fdbserver/logsystem/LogSystem.h"
 #include "fdbserver/logsystem/LogSystemFactory.h"
 #include "fdbserver/core/ServerDBInfo.h"
 #include "fdbserver/core/WaitFailure.h"
