+++ fdbserver/backupworker/BackupWorkerRangePartitioned.cpp
@@ -25,9 +25,9 @@
 #include "fdbclient/SystemData.h"
 #include "fdbclient/Tracing.h"
 #include "BackupPartitionMap.h"
-#include "fdbserver/core/BackupRangePartitionedProgress.h"
+#include "fdbserver/logsystem/BackupRangePartitionedProgress.h"
 #include "fdbserver/core/Knobs.h"
-#include "fdbserver/core/LogSystem.h"
+#include "fdbserver/logsystem/LogSystem.h"
 #include "fdbserver/core/WaitFailure.h"
 #include "fdbserver/logsystem/LogSystemFactory.h"
 #include "flow/CoroUtils.h"
