+++ fdbserver/clustercontroller/ClusterController.actor.cpp
@@ -45,7 +45,7 @@
 #include "ClusterRecovery.actor.h"
 #include "fdbserver/core/DataDistributorInterface.h"
 #include "fdbserver/core/LeaderElection.h"
-#include "fdbserver/core/LogSystem.h"
+#include "fdbserver/logsystem/LogSystem.h"
 #include "fdbserver/core/LogSystemConfig.h"
 #include "fdbserver/logsystem/LogSystemDiskQueueAdapter.h"
 #include "fdbserver/core/WaitFailure.h"
@@ -61,8 +61,8 @@
 #include "fdbrpc/ReplicationUtils.h"
 #include "fdbrpc/sim_validation.h"
 #include "fdbclient/KeyBackedTypes.actor.h"
-#include "fdbserver/core/ApplyMetadataMutation.h"
-#include "fdbserver/core/BackupProgress.h"
+#include "fdbserver/logsystem/ApplyMetadataMutation.h"
+#include "fdbserver/logsystem/BackupProgress.h"
 #include "fdbserver/core/DBCoreState.h"
 #include "fdbserver/core/MoveKeys.h"
 #include "flow/Error.h"
