+++ fdbserver/logsystem/ApplyMetadataMutation.cpp
@@ -24,11 +24,11 @@
 #include "fdbclient/Notified.h"
 #include "fdbclient/SystemData.h"
 #include "fdbserver/core/AccumulativeChecksumUtil.h"
-#include "fdbserver/core/ApplyMetadataMutation.h"
 #include "fdbserver/core/IKeyValueStore.h"
 #include "fdbserver/core/Knobs.h"
 #include "fdbserver/core/LogProtocolMessage.h"
-#include "fdbserver/core/LogSystem.h"
+#include "fdbserver/logsystem/ApplyMetadataMutation.h"
+#include "fdbserver/logsystem/LogSystem.h"
 #include "flow/Error.h"
 #include "flow/Trace.h"
