+++ fdbserver/logsystem/LogSystem.cpp
@@ -19,6 +19,10 @@
  */
 
 #include "fdbserver/logsystem/LogSystem.h"
+#include "fdbclient/FDBTypes.h"
+#include "fdbserver/core/OTELSpanContextMessage.h"
+#include "fdbserver/core/SpanContextMessage.h"
+#include "flow/serialize.h"
 
 bool logSystemHasRemoteLogs(LogSystem const& logSystem) {
 	return logSystem.hasRemoteLogs();
@@ -51,6 +55,123 @@ Tag logSystemGetRandomTxsTag(LogSystem const& logSystem) {
 
 TLogVersion logSystemGetTLogVersion(LogSystem const& logSystem) {
 	return logSystem.getTLogVersion();
+}
+
+LogPushData::LogPushData(Reference<LogSystem> logSystem, int tlogCount) : logSystem(logSystem), subsequence(1) {
+	ASSERT(tlogCount > 0);
+	messagesWriter.reserve(tlogCount);
+	for (int i = 0; i < tlogCount; i++) {
+		messagesWriter.emplace_back(AssumeVersion(g_network->protocolVersion()));
+	}
+	messagesWritten = std::vector<bool>(tlogCount, false);
+}
+
+void LogPushData::addTxsTag() {
+	next_message_tags.push_back(logSystemGetRandomTxsTag(*logSystem));
+}
+
+void LogPushData::addTransactionInfo(SpanContext const& context) {
+	CODE_PROBE(!spanContext.isValid(), "addTransactionInfo with invalid SpanContext");
+	spanContext = context;
+	writtenLocations.clear();
+}
+
+void LogPushData::writeMessage(StringRef rawMessageWithoutLength, bool usePreviousLocations) {
+	if (!usePreviousLocations) {
+		prev_tags.clear();
+		if (logSystemHasRemoteLogs(*logSystem)) {
+			prev_tags.push_back(chooseRouterTag());
+		}
+		for (auto& tag : next_message_tags) {
+			prev_tags.push_back(tag);
+		}
+		msg_locations.clear();
+		logSystemGetPushLocations(*logSystem, VectorRef<Tag>((Tag*)prev_tags.data(), prev_tags.size()), msg_locations);
+		written_tags.insert(next_message_tags.begin(), next_message_tags.end());
+		next_message_tags.clear();
+	}
+	uint32_t subseq = this->subsequence++;
+	uint32_t msgsize =
+	    rawMessageWithoutLength.size() + sizeof(subseq) + sizeof(uint16_t) + sizeof(Tag) * prev_tags.size();
+	for (int loc : msg_locations) {
+		BinaryWriter& wr = messagesWriter[loc];
+		wr << msgsize << subseq << uint16_t(prev_tags.size());
+		for (auto& tag : prev_tags)
+			wr << tag;
+		wr.serializeBytes(rawMessageWithoutLength);
+	}
+}
+
+std::vector<Standalone<StringRef>> LogPushData::getAllMessages() const {
+	std::vector<Standalone<StringRef>> results;
+	results.reserve(messagesWriter.size());
+	for (int loc = 0; loc < messagesWriter.size(); loc++) {
+		results.push_back(getMessages(loc));
+	}
+	return results;
+}
+
+void LogPushData::recordEmptyMessage(int loc, const Standalone<StringRef>& value) {
+	if (!messagesWritten[loc]) {
+		BinaryWriter w(AssumeVersion(g_network->protocolVersion()));
+		Standalone<StringRef> v = w.toValue();
+		if (value.size() > v.size()) {
+			messagesWritten[loc] = true;
+		}
+	}
+}
+
+float LogPushData::getEmptyMessageRatio() const {
+	auto count = std::count(messagesWritten.begin(), messagesWritten.end(), false);
+	ASSERT_WE_THINK(!messagesWritten.empty());
+	return 1.0 * count / messagesWritten.size();
+}
+
+bool LogPushData::writeTransactionInfo(int location, uint32_t subseq) {
+	if (!FLOW_KNOBS->WRITE_TRACING_ENABLED || logSystemGetTLogVersion(*logSystem) < TLogVersion::V6 ||
+	    writtenLocations.contains(location)) {
+		return false;
+	}
+
+	CODE_PROBE(true, "Wrote SpanContextMessage to a transaction log");
+	writtenLocations.insert(location);
+
+	BinaryWriter& wr = messagesWriter[location];
+	int offset = wr.getLength();
+	wr << uint32_t(0) << subseq << uint16_t(prev_tags.size());
+	for (auto& tag : prev_tags)
+		wr << tag;
+	if (logSystemGetTLogVersion(*logSystem) >= TLogVersion::V7) {
+		OTELSpanContextMessage contextMessage(spanContext);
+		wr << contextMessage;
+	} else {
+		SpanContextMessage contextMessage;
+		if (spanContext.isSampled()) {
+			CODE_PROBE(true, "Converting OTELSpanContextMessage to traced SpanContextMessage");
+			contextMessage = SpanContextMessage(UID(spanContext.traceID.first(), spanContext.traceID.second()));
+		} else {
+			CODE_PROBE(true, "Converting OTELSpanContextMessage to untraced SpanContextMessage");
+			contextMessage = SpanContextMessage(UID(0, 0));
+		}
+		wr << contextMessage;
+	}
+	int length = wr.getLength() - offset;
+	*(uint32_t*)((uint8_t*)wr.getData() + offset) = length - sizeof(uint32_t);
+	return true;
+}
+
+void LogPushData::setMutations(uint32_t totalMutations, VectorRef<StringRef> mutations) {
+	ASSERT_EQ(subsequence, 1);
+	subsequence = totalMutations + 1;
+
+	ASSERT_EQ(messagesWriter.size(), mutations.size());
+	BinaryWriter w(AssumeVersion(g_network->protocolVersion()));
+	Standalone<StringRef> v = w.toValue();
+	const int header = v.size();
+	for (int i = 0; i < mutations.size(); i++) {
+		BinaryWriter& wr = messagesWriter[i];
+		wr.serializeBytes(mutations[i].substr(header));
+	}
 }
 
 #include <boost/dynamic_bitset.hpp>
