+++ fdbserver/logsystem/LogSystemDiskQueueAdapter.cpp
@@ -19,7 +19,7 @@
  */
 
 #include "fdbserver/core/IDiskQueue.h"
-#include "fdbserver/core/LogSystem.h"
+#include "fdbserver/logsystem/LogSystem.h"
 #include "fdbserver/logsystem/LogSystemDiskQueueAdapter.h"
 #include "fdbserver/core/Knobs.h"
 #include "flow/CoroUtils.h"
