+++ fdbserver/logsystem/include/fdbserver/logsystem/LogSystem.h
@@ -23,10 +23,199 @@
 
 #pragma once
 
+#include <algorithm>
+#include <cstdint>
+#include <set>
+#include <unordered_set>
+#include <vector>
+
+#include "fdbclient/DatabaseConfiguration.h"
 #include "fdbrpc/Replication.h"
+#include "fdbrpc/Locality.h"
+#include "fdbrpc/ReplicationPolicy.h"
 #include "fdbserver/core/DBCoreState.h"
-#include "LogSystemTypes.h"
+#include "fdbserver/core/LogSystemConfig.h"
+#include "fdbserver/core/MutationTracking.h"
+#include "fdbserver/core/OTELSpanContextMessage.h"
+#include "fdbserver/core/SpanContextMessage.h"
+#include "fdbserver/core/TLogInterface.h"
+#include "fdbserver/core/WorkerInterface.actor.h"
+#include "flow/Arena.h"
+#include "flow/Error.h"
 #include "flow/ActorCollection.h"
+#include "flow/Histogram.h"
+#include "flow/IndexedSet.h"
+#include "flow/Knobs.h"
+
+struct DBCoreState;
+struct LogPushData;
+struct LocalityData;
+struct LogSystem;
+class LogSet;
+
+struct ConnectionResetInfo : public ReferenceCounted<ConnectionResetInfo> {
+	double lastReset;
+	Future<Void> resetCheck;
+	int slowReplies;
+	int fastReplies;
+
+	ConnectionResetInfo() : lastReset(now()), resetCheck(Void()), slowReplies(0), fastReplies(0) {}
+};
+
+struct IPeekCursor {
+	virtual Reference<IPeekCursor> cloneNoMore() = 0;
+
+	virtual void setProtocolVersion(ProtocolVersion version) = 0;
+
+	virtual bool hasMessage() const = 0;
+	virtual VectorRef<Tag> getTags() const = 0;
+	virtual Arena& arena() = 0;
+	virtual ArenaReader* reader() = 0;
+	virtual StringRef getMessage() = 0;
+	virtual StringRef getMessageWithTags() = 0;
+	virtual void nextMessage() = 0;
+	virtual void advanceTo(LogMessageVersion n) = 0;
+	virtual Future<Void> getMore(TaskPriority taskID = TaskPriority::TLogPeekReply) = 0;
+	virtual Future<Void> onFailed() const = 0;
+	virtual bool isActive() const = 0;
+	virtual bool isExhausted() const = 0;
+	virtual const LogMessageVersion& version() const = 0;
+	virtual Version popped() const = 0;
+	virtual Version getMaxKnownVersion() const { return 0; }
+	virtual Version getMinKnownCommittedVersion() const = 0;
+	virtual Optional<UID> getPrimaryPeekLocation() const = 0;
+	virtual Optional<UID> getCurrentPeekLocation() const = 0;
+	virtual void addref() = 0;
+	virtual void delref() = 0;
+};
+
+struct LogPushVersionSet {
+	Version prevVersion;
+	Version version;
+	Version knownCommittedVersion;
+	Version minKnownCommittedVersion;
+};
+
+struct EpochTagsVersionsInfo {
+	int32_t logRouterTags;
+	Version epochBegin, epochEnd;
+
+	explicit EpochTagsVersionsInfo(int32_t n, Version begin, Version end)
+	  : logRouterTags(n), epochBegin(begin), epochEnd(end) {}
+};
+
+bool logSystemHasRemoteLogs(LogSystem const& logSystem);
+void logSystemGetPushLocations(
+    LogSystem const& logSystem,
+    VectorRef<Tag> tags,
+    std::vector<int>& locations,
+    bool allLocations = false,
+    Optional<std::vector<Reference<LocalitySet>>> fromLocations = Optional<std::vector<Reference<LocalitySet>>>());
+std::vector<Reference<LocalitySet>> logSystemGetPushLocationsForTags(LogSystem const& logSystem,
+                                                                     std::vector<int>& fromLocations);
+Tag logSystemGetRandomRouterTag(LogSystem const& logSystem);
+int logSystemGetLogRouterTags(LogSystem const& logSystem);
+Tag logSystemGetRandomTxsTag(LogSystem const& logSystem);
+TLogVersion logSystemGetTLogVersion(LogSystem const& logSystem);
+
+struct LengthPrefixedStringRef {
+	uint32_t* length;
+
+	StringRef toStringRef() const {
+		ASSERT(length);
+		return StringRef((uint8_t*)(length + 1), *length);
+	}
+	int expectedSize() const {
+		ASSERT(length);
+		return *length;
+	}
+	uint32_t* getLengthPtr() const { return length; }
+
+	LengthPrefixedStringRef() : length(nullptr) {}
+	LengthPrefixedStringRef(uint32_t* length) : length(length) {}
+};
+
+struct LogPushData : NonCopyable {
+	explicit LogPushData(Reference<LogSystem> logSystem, int tlogCount);
+
+	void addTxsTag();
+
+	void addTag(Tag tag) { next_message_tags.push_back(tag); }
+
+	template <class T>
+	void addTags(T tags) {
+		next_message_tags.insert(next_message_tags.end(), tags.begin(), tags.end());
+	}
+
+	void addTransactionInfo(SpanContext const& context);
+
+	void saveTags(std::set<Tag>& filteredTags) const {
+		for (const auto& tag : written_tags) {
+			filteredTags.insert(tag);
+		}
+	}
+
+	void addWrittenTags(const std::set<Tag>& tags) { written_tags.insert(tags.begin(), tags.end()); }
+
+	void getLocations(const std::set<Tag>& tags, std::set<uint16_t>& writtenTLogs) {
+		std::vector<Tag> vtags(tags.begin(), tags.end());
+		std::vector<int> locations;
+		logSystemGetPushLocations(*logSystem, VectorRef<Tag>((Tag*)vtags.data(), vtags.size()), locations);
+		writtenTLogs.insert(locations.begin(), locations.end());
+	}
+
+	void saveLocations(std::set<uint16_t>& writtenTLogs) {
+		writtenTLogs.insert(msg_locations.begin(), msg_locations.end());
+	}
+
+	void setLogsChanged() { logsChanged = true; }
+	bool haveLogsChanged() const { return logsChanged; }
+
+	void writeMessage(StringRef rawMessageWithoutLength, bool usePreviousLocations);
+
+	void setPushLocationsForTags(std::vector<int> fromLocationsVec) {
+		fromLocations = logSystemGetPushLocationsForTags(*logSystem, fromLocationsVec);
+	}
+
+	template <class T>
+	void writeTypedMessage(T const& item, bool metadataMessage = false, bool allLocations = false);
+
+	Standalone<StringRef> getMessages(int loc) const { return messagesWriter[loc].toValue(); }
+
+	std::vector<Standalone<StringRef>> getAllMessages() const;
+
+	void recordEmptyMessage(int loc, const Standalone<StringRef>& value);
+
+	float getEmptyMessageRatio() const;
+
+	uint32_t getMutationCount() const { return subsequence - 1; }
+
+	void setMutations(uint32_t totalMutations, VectorRef<StringRef> mutations);
+
+	Optional<Tag> savedRandomRouterTag;
+	void storeRandomRouterTag() { savedRandomRouterTag = logSystemGetRandomRouterTag(*logSystem); }
+	int getLogRouterTags() { return logSystemGetLogRouterTags(*logSystem); }
+
+private:
+	Reference<LogSystem> logSystem;
+	std::vector<Tag> next_message_tags;
+	std::vector<Tag> prev_tags;
+	std::set<Tag> written_tags;
+	std::vector<BinaryWriter> messagesWriter;
+	std::vector<bool> messagesWritten;
+	std::vector<int> msg_locations;
+	Optional<std::vector<Reference<LocalitySet>>> fromLocations;
+	std::unordered_set<int> writtenLocations;
+	uint32_t subsequence;
+	SpanContext spanContext;
+	bool logsChanged = false;
+
+	bool writeTransactionInfo(int location, uint32_t subseq);
+
+	Tag chooseRouterTag() {
+		return savedRandomRouterTag.present() ? savedRandomRouterTag.get() : logSystemGetRandomRouterTag(*logSystem);
+	}
+};
 
 // LogSystem info in old epoch
 struct OldLogData {
@@ -42,15 +231,7 @@ struct OldLogData {
 
 	// Constructor for T of OldTLogConf and OldTLogCoreData
 	template <class T>
-	explicit OldLogData(const T& conf)
-	  : logRouterTags(conf.logRouterTags), txsTags(conf.txsTags), epochBegin(conf.epochBegin), epochEnd(conf.epochEnd),
-	    recoverAt(conf.recoverAt), pseudoLocalities(conf.pseudoLocalities), epoch(conf.epoch) {
-		tLogs.resize(conf.tLogs.size());
-		for (int j = 0; j < conf.tLogs.size(); j++) {
-			auto logSet = makeReference<LogSet>(conf.tLogs[j]);
-			tLogs[j] = logSet;
-		}
-	}
+	explicit OldLogData(const T& conf);
 };
 
 struct IdToInterf : ReferenceCounted<IdToInterf> {
@@ -440,6 +621,79 @@ std::vector<T> LogSystem::getReadyNonError(std::vector<Future<T>> const& futures
 		if (f.isReady() && !f.isError())
 			result.push_back(f.get());
 	return result;
+}
+
+#include "LogSystemTypes.h"
+
+template <class T>
+OldLogData::OldLogData(const T& conf)
+  : logRouterTags(conf.logRouterTags), txsTags(conf.txsTags), epochBegin(conf.epochBegin), epochEnd(conf.epochEnd),
+    recoverAt(conf.recoverAt), pseudoLocalities(conf.pseudoLocalities), epoch(conf.epoch) {
+	tLogs.resize(conf.tLogs.size());
+	for (int j = 0; j < conf.tLogs.size(); j++) {
+		auto logSet = makeReference<LogSet>(conf.tLogs[j]);
+		tLogs[j] = logSet;
+	}
+}
+
+template <class T>
+void LogPushData::writeTypedMessage(T const& item, bool metadataMessage, bool allLocations) {
+	prev_tags.clear();
+	if (logSystemHasRemoteLogs(*logSystem)) {
+		prev_tags.push_back(chooseRouterTag());
+	}
+	for (auto& tag : next_message_tags) {
+		prev_tags.push_back(tag);
+	}
+	msg_locations.clear();
+	logSystemGetPushLocations(*logSystem,
+	                          VectorRef<Tag>((Tag*)prev_tags.data(), prev_tags.size()),
+	                          msg_locations,
+	                          allLocations,
+	                          fromLocations);
+
+	BinaryWriter bw(AssumeVersion(g_network->protocolVersion()));
+
+	if (!metadataMessage) {
+		uint32_t subseq = this->subsequence++;
+		bool updatedLocation = false;
+		for (int loc : msg_locations) {
+			updatedLocation = writeTransactionInfo(loc, subseq) || updatedLocation;
+		}
+		if (!updatedLocation) {
+			this->subsequence--;
+			CODE_PROBE(true, "No new SpanContextMessage written to transaction logs");
+			ASSERT(this->subsequence > 0);
+		}
+	} else {
+		ASSERT(writtenLocations.size() == 0);
+	}
+
+	uint32_t subseq = this->subsequence++;
+	bool first = true;
+	int firstOffset = -1, firstLength = -1;
+	for (int loc : msg_locations) {
+		BinaryWriter& wr = messagesWriter[loc];
+
+		if (first) {
+			firstOffset = wr.getLength();
+			wr << uint32_t(0) << subseq << uint16_t(prev_tags.size());
+			for (auto& tag : prev_tags)
+				wr << tag;
+			wr << item;
+			firstLength = wr.getLength() - firstOffset;
+			*(uint32_t*)((uint8_t*)wr.getData() + firstOffset) = firstLength - sizeof(uint32_t);
+			DEBUG_TAGS_AND_MESSAGE(
+			    "ProxyPushLocations", invalidVersion, StringRef(((uint8_t*)wr.getData() + firstOffset), firstLength))
+			    .detail("PushLocations", msg_locations);
+			first = false;
+		} else {
+			BinaryWriter& from = messagesWriter[msg_locations[0]];
+			wr.serializeBytes((uint8_t*)from.getData() + firstOffset, firstLength);
+		}
+	}
+	written_tags.insert(next_message_tags.begin(), next_message_tags.end());
+	next_message_tags.clear();
 }
 
 #endif // FDBSERVER_LOGSYSTEM_LOGSYSTEM_H
