+++ fdbserver/tlog/TestTLogServer.cpp
@@ -31,7 +31,7 @@
 #include "fdbserver/core/TLogInterface.h"
 #include "fdbserver/tlog/TLogServer.actor.h"
 #include "fdbserver/core/WorkerInterface.actor.h"
-#include "fdbserver/core/LogSystem.h"
+#include "fdbserver/logsystem/LogSystem.h"
 #include "fdbserver/logsystem/LogSystemFactory.h"
 #include "flow/IRandom.h"
 #include "flow/DebugTrace.h"
